+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Romantic Versioning](http://dafoster.net/articles/2
 - Zsh environment related settings now confirmed into one file
 - Zsh history is only kept for non-privileged user accounts
 - Zsh alias for GNU tr instead of BSD tr
+- Zsh alias update! command is now a function and includes gem and pip updates
 
 ### Deprecated
 - Soon-to-be removed feature